Move test-specific code into tests and fix fraud proof test
